Cast slice length to size_t

+++ src/mtype.c
@@ -9295,7 +9295,7 @@ Type *TypeSlice::semantic(Loc loc, Scope *sc)
     }
 
     Parameters *args = new Parameters;
-    args->reserve(i2 - i1);
+    args->reserve((size_t)(i2 - i1));
     for (size_t i = i1; i < i2; i++)
     {   Parameter *arg = (*tt->arguments)[i];
         args->push(arg);
